requirement.txt
discord.py == 2.5.2
python-dotenv